Correctly define LONG_PTR, ULONG_PTR and SIZE_T in "typedefs_host.h". This should silence many warnings and might fix the Live-CD build problems of our 64-bit Buildslave.
svn path=/trunk/; revision=28844

+++ reactos/tools/mkhive/reginf.c
@@ -181,7 +181,7 @@ do_reg_operation(
 {
 	CHAR EmptyStr = (CHAR)0;
 	ULONG Type;
-	SIZE_T Size;
+	ULONG Size;
 	LONG Error;
 
 	if (Flags & FLG_ADDREG_DELVAL)  /* deletion */
